refactor: SpotifyService 리팩토링
