+++ src/main/java/com/back/web7_9_codecrete_be/domain/artists/service/spotify/genre/CategoryConfig.java
@@ -1,9 +1,8 @@
-package com.back.web7_9_codecrete_be.domain.artists.service.spotifyService;
+package com.back.web7_9_codecrete_be.domain.artists.service.spotify.genre;
 
 import java.util.List;
 
-// 카테고리별 키워드 및 상한 설정
-
+// 카테고리별 키워드 및 상한 설정 : 장르/카테고리별 수집 규칙 정의
 public class CategoryConfig {
     public final List<String> keywords;
     public final int targetCount;
